+++ source/app.py
@@ -1,41 +1,21 @@
 from starlette.applications import Starlette
 from starlette.exceptions import HTTPException
-from starlette.staticfiles import StaticFiles
 from starlette.responses import HTMLResponse, RedirectResponse
-from starlette.templating import Jinja2Templates
 from sentry_sdk.integrations.asgi import SentryAsgiMiddleware
 from source import settings, pagination, ordering, search, tables
+from source.resources import database, statics, templates
+from source.datasource import ElectionDataSource
 import databases
-import sentry_sdk
 import math
 import typesystem
 
-
-if settings.SENTRY_DSN:  # pragma: nocover
-    sentry_sdk.init(dsn=settings.SENTRY_DSN, release=settings.RELEASE_VERSION)
-
-
-templates = Jinja2Templates(directory="templates")
-
-if settings.TESTING:
-    database = databases.Database(settings.TEST_DATABASE_URL, force_rollback=True)
-else:  # pragma: nocover
-    database = databases.Database(settings.DATABASE_URL)
 
 app = Starlette(debug=settings.DEBUG)
 
 if settings.SENTRY_DSN:  # pragma: nocover
     app.add_middleware(SentryAsgiMiddleware)
 
-app.mount("/static", StaticFiles(directory="statics"), name="static")
-
-
-class Record(typesystem.Schema):
-    constituency = typesystem.String(title="Constituency", max_length=100)
-    surname = typesystem.String(title="Surname", max_length=100)
-    first_name = typesystem.String(title="First Name", max_length=100)
-    party = typesystem.String(title="Party", max_length=100)
-    votes = typesystem.Integer(title="Votes", minimum=0)
+app.mount("/static", statics, name="static")
 
 
 @app.on_event("startup")
@@ -46,137 +26,6 @@ async def startup():
 @app.on_event("shutdown")
 async def shutdown():
     await database.disconnect()
-
-
-class ElectionDataSource:
-    schema = Record
-
-    def __init__(self, app, year):
-        self.name = f"UK General Election Results {year}"
-        self.app = app
-        self.url = app.url_path_for("table", year=year)
-        self.year = year
-        self.clauses = []
-        self.order_column = None
-        self.query_limit = None
-        self.query_offset = None
-
-    def apply_query_filters(self, query):
-        query = query.where(tables.election.c.year == self.year)
-
-        for clause in self.clauses:
-            query = query.where(clause)
-
-        if self.query_limit is not None:
-            query = query.limit(self.query_limit)
-
-        if self.query_offset is not None:
-            query = query.offset(self.query_offset)
-
-        if self.order_column is not None:
-            query = query.group_by(tables.election.c.pk).order_by(
-                self.order_column, tables.election.c.pk
-            )
-        return query
-
-    def limit(self, limit):
-        self.query_limit = limit
-        return self
-
-    def offset(self, offset):
-        self.query_offset = offset
-        return self
-
-    def search(self, search_term):
-        if not search_term:
-            return self
-
-        match = f"%{search_term}%"
-        self.clauses.append(
-            (
-                tables.election.c.constituency.ilike(match)
-                | tables.election.c.surname.ilike(match)
-                | tables.election.c.first_name.ilike(match)
-                | tables.election.c.party.ilike(match)
-            )
-        )
-        return self
-
-    def filter(self, pk=None):
-        self.clauses.append(tables.election.c.pk == pk)
-        return self
-
-    def order_by(self, column, reverse):
-        order_column = {
-            "constituency": tables.election.c.constituency,
-            "surname": tables.election.c.surname,
-            "first_name": tables.election.c.first_name,
-            "party": tables.election.c.party,
-            "votes": tables.election.c.votes,
-        }[column]
-        self.order_column = order_column.desc() if reverse else order_column.asc()
-        return self
-
-    async def count(self):
-        query = tables.election.count()
-        query = self.apply_query_filters(query)
-        return await database.fetch_val(query)
-
-    async def all(self):
-        query = tables.election.select()
-        query = self.apply_query_filters(query)
-        rows = await database.fetch_all(query)
-        return [ElectionDataItem(app=self.app, year=self.year, row=row) for row in rows]
-
-    async def get(self):
-        query = tables.election.select()
-        query = self.apply_query_filters(query)
-        row = await database.fetch_one(query)
-        if row is None:
-            return None
-        return ElectionDataItem(app=self.app, year=self.year, row=row)
-
-    async def create(self, values):
-        values = dict(values)
-        values["year"] = self.year
-        query = tables.election.insert()
-        return await database.execute(query, values=values)
-
-    def validate(self, data):
-        record, errors = Record.validate_or_error(data)
-        validated_data = dict(record) if record is not None else None
-        return validated_data, errors
-
-
-class ElectionDataItem:
-    def __init__(self, app, year, row):
-        self.app = app
-        self.year = year
-        self.pk = row["pk"]
-        self.constituency = row["constituency"]
-        self.surname = row["surname"]
-        self.first_name = row["first_name"]
-        self.party = row["party"]
-        self.votes = row["votes"]
-
-    def __str__(self):
-        return f"{self.first_name} {self.surname}"
-
-    async def update(self, values):
-        query = tables.election.update().where(tables.election.c.pk == self.pk)
-        return await database.execute(query, values=values)
-
-    async def delete(self):
-        query = tables.election.delete().where(tables.election.c.pk == self.pk)
-        return await database.execute(query)
-
-    @property
-    def url(self):
-        return self.app.url_path_for("detail", year=self.year, pk=self.pk)
-
-    @property
-    def delete_url(self):
-        return self.app.url_path_for("delete-row", year=self.year, pk=self.pk)
 
 
 @app.route("/", name="dashboard")
